rmcp

Attribute Macro prompt

#[prompt]

prompt

This macro is used to mark a function as a prompt handler.

This will generate a function that returns the attribute of this prompt, with type rmcp::model::Prompt.

Usage

field	type	usage

name	String	The name of the prompt. If not provided, it defaults to the function name.

description	String	A description of the prompt. The document of this function will be used if not provided.

arguments	Expr	An expression that evaluates to Option<Vec<PromptArgument>> defining the prompt’s arguments. If not provided, it will automatically generate arguments from the Parameters<T> type found in the function signature.

Example

#[prompt(name = "code_review", description = "Reviews code for best practices")]
pub async fn code_review_prompt(&self, Parameters(args): Parameters<CodeReviewArgs>) -> Result<Vec<PromptMessage>> {
    // Generate prompt messages based on arguments
}