rmcp::model

Module content

Content sent around agents, extensions, and LLMs The various content types can be display to humans but also understood by models They include optional annotations used to help inform agent usage

Structs

JsonContent

RawAudioContent

RawEmbeddedResource

RawImageContent

RawTextContent

ToolResultContent

Tool execution result in user message (SEP-1577).

ToolUseContent

Tool call request from assistant (SEP-1577).

Enums

RawContent

Traits

IntoContents

Types that can be converted into a list of contents

Type Aliases

AudioContent

Content

EmbeddedResource

ImageContent

TextContent