rmcp::model

Struct EnumSchemaBuilder

pub struct EnumSchemaBuilder<T> { /* private fields */ }

Builder for EnumSchema Allows to create various enum schema types (single/multi select, titled/untitled) with validation of provided parameters

Example

use rmcp::model::*;
let enum_schema = EnumSchema::builder(vec!["Red".to_string(), "Green".to_string(), "Blue".to_string()])
 .multiselect()
 .enum_titles(vec!["Red Color".to_string(), "Green Color".to_string(), "Blue Color".to_string()])
 .expect("Number of titles should match number of values")
 .min_items(1u64).expect("Min items should be correct value")
 .max_items(3u64).expect("Max items should be correct value")
 .description("Select your favorite colors")
 .build();

Implementations

impl<T> EnumSchemaBuilder<T>

Common enum schema builder methods

pub fn title(self, value: impl Into<Cow<'static, str>>) -> Self

Set title of enum schema

pub fn description(self, value: impl Into<Cow<'static, str>>) -> Self

Set description of enum schema

pub fn untitled(self) -> Self

Set enum as untitled Clears any previously set titles

pub fn enum_titles( self, titles: Vec<String>, ) -> Result<EnumSchemaBuilder<T>, String>

Set titles to enum values. Also, implicitly set this enum schema as titled

impl EnumSchemaBuilder<SingleSelect>

Enum selection builder for single-select enums

pub fn new(values: Vec<String>) -> EnumSchemaBuilder<SingleSelect>

pub fn multiselect(self) -> EnumSchemaBuilder<MultiSelect>

Transition to multi-select enum builder.

Clears any previously set default values and resets min/max items. After this transition, you can use min_items(), max_items(), and with_default() for multi-select semantics.

pub fn with_default( self, default_value: impl Into<String>, ) -> Result<EnumSchemaBuilder<SingleSelect>, String>

Set default value

pub fn build(self) -> EnumSchema

Build enum schema

impl EnumSchemaBuilder<MultiSelect>

Enum selection builder for multi-select enums

pub fn single_select(self) -> EnumSchemaBuilder<SingleSelect>

Set enum as single-select If it was multi-select, clear default values

pub fn with_default( self, default_values: Vec<String>, ) -> Result<EnumSchemaBuilder<MultiSelect>, String>

Set default values

pub fn min_items( self, value: u64, ) -> Result<EnumSchemaBuilder<MultiSelect>, String>

Set minimal number of items for multi-select enum options

pub fn max_items( self, value: u64, ) -> Result<EnumSchemaBuilder<MultiSelect>, String>

Set maximal number of items for multi-select enum options

pub fn build(self) -> EnumSchema

Build enum schema