rmcp::model

Struct Notification

#[non_exhaustive]
pub struct Notification<M = String, P = JsonObject> {
    pub method: M,
    pub params: P,
    pub extensions: Extensions,
}

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

method: Mparams: Pextensions: Extensions

extensions will carry anything possible in the context, including Meta

this is similar with the Extensions in http crate

Implementations

impl<M: Default, P> Notification<M, P>

pub fn new(params: P) -> Self