rmcp::model

Struct ToolAnnotations

pub struct ToolAnnotations {
    pub title: Option<String>,
    pub read_only_hint: Option<bool>,
    pub destructive_hint: Option<bool>,
    pub idempotent_hint: Option<bool>,
    pub open_world_hint: Option<bool>,
}

Additional properties describing a Tool to clients.

NOTE: all properties in ToolAnnotations are hints. They are not guaranteed to provide a faithful description of tool behavior (including descriptive properties like title).

Clients should never make tool use decisions based on ToolAnnotations received from untrusted servers.

Fields
title: Option<String>

A human-readable title for the tool.

read_only_hint: Option<bool>

If true, the tool does not modify its environment.

Default: false

destructive_hint: Option<bool>

If true, the tool may perform destructive updates to its environment. If false, the tool performs only additive updates.

(This property is meaningful only when readOnlyHint == false)

Default: true A human-readable description of the tool’s purpose.

idempotent_hint: Option<bool>

If true, calling the tool repeatedly with the same arguments will have no additional effect on the its environment.

(This property is meaningful only when readOnlyHint == false)

Default: false.

open_world_hint: Option<bool>

If true, this tool may interact with an “open world” of external entities. If false, the tool’s domain of interaction is closed. For example, the world of a web search tool is open, whereas that of a memory tool is not.

Default: true

Implementations

impl ToolAnnotations

pub fn new() -> Self

pub fn with_title<T>(title: T) -> Self
where T: Into<String>,

pub fn read_only(self, read_only: bool) -> Self

pub fn destructive(self, destructive: bool) -> Self

pub fn idempotent(self, idempotent: bool) -> Self

pub fn open_world(self, open_world: bool) -> Self

pub fn is_destructive(&self) -> bool

If not set, defaults to true.

pub fn is_idempotent(&self) -> bool

If not set, defaults to false.