rmcp::model

Type Alias CancelledNotification

pub type CancelledNotification = Notification<CancelledNotificationMethod, CancelledNotificationParam>;

Cancellation

This notification can be sent by either side to indicate that it is cancelling a previously-issued request.

The request SHOULD still be in-flight, but due to communication latency, it is always possible that this notification MAY arrive after the request has already finished.

This notification indicates that the result will be unused, so any associated processing SHOULD cease.

A client MUST NOT attempt to cancel its initialize request.

Aliased Type

struct CancelledNotification {
    pub method: CancelledNotificationMethod,
    pub params: CancelledNotificationParam,
}

Fields
method: CancelledNotificationMethodparams: CancelledNotificationParam