rmcp::model

Type Alias ProgressNotification

pub type ProgressNotification = Notification<ProgressNotificationMethod, ProgressNotificationParam>;

Aliased Type

pub struct ProgressNotification {
    pub method: ProgressNotificationMethod,
    pub params: ProgressNotificationParam,
    pub extensions: Extensions,
}

Fields
method: ProgressNotificationMethodparams: ProgressNotificationParamextensions: Extensions

extensions will carry anything possible in the context, including Meta

this is similar with the Extensions in http crate