rmcp

Module service

Structs

AtomicU32Provider

NotificationContext

Request execution context

Peer

An interface to fetch the remote client or server

PeerRequestOptions

RequestContext

Request execution context

RequestHandle

A handle to a remote request

RoleServer

RunningService

RunningServiceCancellationToken

Enums

QuitReason

ServerInitializeError

It represents the error that may occur when serving the server.

ServiceError

Traits

DynService

ProgressTokenProvider

RequestIdProvider

Service

ServiceExt

ServiceRole

Functions

serve_directly

Use this function to skip initialization process

serve_directly_with_ct

Use this function to skip initialization process

serve_server

serve_server_with_ct

Type Aliases

AtomicU32ProgressTokenProvider

AtomicU32RequestIdProvider

ClientSink

RxJsonRpcMessage

TxJsonRpcMessage