rmcp

Module model

Basic data types in MCP specification

Structs

Annotated

Annotations

ArgumentInfo

CallToolRequestMethod

CallToolRequestParam

CallToolResult

CancelledNotificationMethod

CancelledNotificationParam

ClientCapabilities

Builder

ClientCapabilitiesBuilder

ClientCapabilitiesBuilderState

CompleteRequestMethod

CompleteRequestParam

CompleteResult

CompletionInfo

CreateMessageRequestMethod

CreateMessageRequestParam

CreateMessageResult

EmptyObject

ErrorCode

ErrorData

Error information for JSON-RPC error responses.

GetPromptRequestMethod

GetPromptRequestParam

GetPromptResult

Implementation

InitializeRequestParam

InitializeResult

InitializeResultMethod

InitializedNotificationMethod

JsonContent

JsonRpcError

JsonRpcNotification

JsonRpcRequest

JsonRpcResponse

JsonRpcVersion2_0

ListPromptsRequestMethod

ListPromptsResult

ListResourceTemplatesRequestMethod

ListResourceTemplatesResult

ListResourcesRequestMethod

ListResourcesResult

ListRootsRequestMethod

ListRootsResult

ListToolsRequestMethod

ListToolsResult

LoggingMessageNotificationMethod

LoggingMessageNotificationParam

ModelHint

ModelPreferences

Notification

NotificationNoParam

PaginatedRequestParamInner

PingRequestMethod

ProgressNotificationMethod

ProgressNotificationParam

Prompt

A prompt that can be used to generate text from a model

PromptArgument

Represents a prompt argument that can be passed to customize the prompt

PromptArgumentTemplate

A template for a prompt argument, this should be identical to PromptArgument

PromptListChangedNotificationMethod

PromptMessage

A message in a prompt conversation

PromptReference

PromptTemplate

A template for a prompt

PromptsCapability

ProtocolVersion

RawEmbeddedResource

RawImageContent

RawResource

Represents a resource in the extension with metadata

RawResourceTemplate

RawTextContent

ReadResourceRequestMethod

ReadResourceRequestParam

ReadResourceResult

Request

RequestMeta

RequestNoParam

ResourceListChangedNotificationMethod

ResourceReference

ResourceUpdatedNotificationMethod

ResourceUpdatedNotificationParam

ResourcesCapability

Root

RootsCapabilities

RootsListChangedNotificationMethod

SamplingMessage

ServerCapabilities

Builder

ServerCapabilitiesBuilder

ServerCapabilitiesBuilderState

SetLevelRequestMethod

SetLevelRequestParam

SubscribeRequestMethod

SubscribeRequestParam

Tool

A tool that can be used by a model.

ToolListChangedNotificationMethod

ToolsCapability

UnsubscribeRequestMethod

UnsubscribeRequestParam

WithMeta

Enums

ClientNotification

ClientRequest

ClientResult

JsonRpcMessage

LoggingLevel

Message

NumberOrString

PromptMessageContent

Content types that can be included in prompt messages

PromptMessageRole

Represents the role of a message sender in a prompt conversation

RawContent

Reference

ResourceContents

Role

ServerNotification

ServerRequest

ServerResult

Traits

AnnotateAble

ConstString

IntoContents

Types that can be converted into a list of contents

Functions

object

unwrap the JsonObject under serde_json::Value

Type Aliases

CallToolRequest

CancelledNotification

Cancellation

ClientInfo

ClientJsonRpcMessage

ClientMessage

CompleteRequest

Content

CreateMessageRequest

Cursor

EmbeddedResource

EmptyResult

Empty result

ExperimentalCapabilities

GetPromptRequest

ImageContent

InitializeRequest

Initialization

InitializedNotification

This notification is sent from the client to the server after initialization has finished.

JsonObject

ListPromptsRequest

ListResourceTemplatesRequest

ListResourcesRequest

ListRootsRequest

ListToolsRequest

LoggingMessageNotification

PaginatedRequestParam

PingRequest

ProgressNotification

ProgressToken

PromptListChangedNotification

ReadResourceRequest

RequestId

Resource

ResourceListChangedNotification

ResourceTemplate

ResourceUpdatedNotification

RootsListChangedNotification

ServerInfo

ServerJsonRpcMessage

ServerMessage

SetLevelRequest

SubscribeRequest

TextContent

ToolListChangedNotification

UnsubscribeRequest