rmcp

Module transport

Transport

The transport type must implemented IntoTransport trait, which allow split into a sink and a stream.

For client, the sink item is ClientJsonRpcMessage and stream item is ServerJsonRpcMessage

For server, the sink item is ServerJsonRpcMessage and stream item is ClientJsonRpcMessage

These types is automatically implemented IntoTransport trait

For type that already implement both Sink and Stream trait, they are automatically implemented IntoTransport trait

For tuple of sink Tx and stream Rx, type (Tx, Rx) are automatically implemented IntoTransport trait

For type that implement both tokio::io::AsyncRead and tokio::io::AsyncWrite trait, they are automatically implemented IntoTransport trait

For tuple of tokio::io::AsyncRead R and tokio::io::AsyncWrite W, type (R, W) are automatically implemented IntoTransport trait

Examples

// create transport from tcp stream
async fn client() -> Result<(), Box<dyn std::error::Error>> {
    let stream = tokio::net::TcpSocket::new_v4()?
        .connect("127.0.0.1:8001".parse()?)
        .await?;
    let client = ().serve(stream).await?;
    let tools = client.peer().list_tools(Default::default()).await?;
    println!("{:?}", tools);
    Ok(())
}

// create transport from std io
async fn io()  -> Result<(), Box<dyn std::error::Error>> {
    let client = None.serve((tokio::io::stdin(), tokio::io::stdout())).await?;
    let tools = client.peer().list_tools(Default::default()).await?;
    println!("{:?}", tools);
    Ok(())
}

Modules

io

Enums

TransportAdapterStreamSink

Traits

IntoTransport