rmcp

Module transport

Transport

The transport type must implemented Transport trait, which allow it send message concurrently and receive message sequentially.

Standard Transport Types

There are 2 pairs of standard transport types:

transport	client	server

std IO	[child_process::TokioChildProcess]	io::stdio

streamable http	[streamable_http_client::StreamableHttpClientTransport]	streamable_http_server::StreamableHttpService

Thers are several helper transport types that can help you to create transport quickly.

Worker Transport

Which allows you to run a worker and process messages in another tokio task.

Async Read/Write Transport

You need to enable transport-async-rw feature to use this transport.

This transport is used to create a transport from a byte stream which implemented [tokio::io::AsyncRead] and [tokio::io::AsyncWrite].

This could be very helpful when you want to create a transport from a byte stream, such as a file or a tcp connection.

Sink/Stream Transport

This transport is used to create a transport from a sink and a stream.

This could be very helpful when you want to create a transport from a duplex object stream, such as a websocket connection.

IntoTransport trait

IntoTransport is a helper trait that implicitly convert a type into a transport type.

These types is automatically implemented IntoTransport trait

A type that already implement both [futures::Sink] and [futures::Stream] trait, or a tuple (Tx, Rx) where Tx is [futures::Sink] and Rx is [futures::Stream].

A type that implement both [tokio::io::AsyncRead] and [tokio::io::AsyncWrite] trait. or a tuple (R, W) where R is [tokio::io::AsyncRead] and W is [tokio::io::AsyncWrite].

A type that implement Worker trait.

A type that implement Transport trait.

Examples

#[cfg(feature = "client")]

// create transport from tcp stream
#[cfg(feature = "client")]
async fn client() -> Result<(), Box<dyn std::error::Error>> {
    let stream = tokio::net::TcpSocket::new_v4()?
        .connect("127.0.0.1:8001".parse()?)
        .await?;
    let client = ().serve(stream).await?;
    let tools = client.peer().list_tools(Default::default()).await?;
    println!("{:?}", tools);
    Ok(())
}

// create transport from std io
#[cfg(feature = "client")]
async fn io()  -> Result<(), Box<dyn std::error::Error>> {
    let client = ().serve((tokio::io::stdin(), tokio::io::stdout())).await?;
    let tools = client.peer().list_tools(Default::default()).await?;
    println!("{:?}", tools);
    Ok(())
}

Re-exports

pub use io::stdio;

Modules

async_rw

common

Common use codes

io

sink_stream

Structs

DynamicTransportError

OneshotTransport

A transport that can send a single message and then close itself

Enums

TransportAdapterIdentity

Traits

IntoTransport

Transport