rmcp::transport

Trait Transport

pub trait Transport<R>: Send
where
    R: ServiceRole,
{
    type Error: Error + Send + Sync + 'static;

    // Required methods
    fn send(
        &mut self,
        item: TxJsonRpcMessage<R>,
    ) -> impl Future<Output = Result<(), Self::Error>> + Send + 'static;
    fn receive(
        &mut self,
    ) -> impl Future<Output = Option<RxJsonRpcMessage<R>>> + Send;
    fn close(&mut self) -> impl Future<Output = Result<(), Self::Error>> + Send;

    // Provided method
    fn name() -> Cow<'static, str> { ... }
}

Required Associated Types

type Error: Error + Send + Sync + 'static

Required Methods

fn send( &mut self, item: TxJsonRpcMessage<R>, ) -> impl Future<Output = Result<(), Self::Error>> + Send + 'static

Send a message to the transport

Notice that the future returned by this function should be Send and 'static. It’s because the sending message could be executed concurrently.

fn receive( &mut self, ) -> impl Future<Output = Option<RxJsonRpcMessage<R>>> + Send

Receive a message from the transport, this operation is sequential.

fn close(&mut self) -> impl Future<Output = Result<(), Self::Error>> + Send

Close the transport

Provided Methods

fn name() -> Cow<'static, str>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.